Move get_kernel_version_list to payload base
This function is used in live_os and live_image so move it to
modules.payload.base.utils python module.
